add `media/interact` tests
